Export sketch components, move them out of /page dir
