+++ src/components/index.ts
@@ -1,0 +1,2 @@
+export { default as FullPageSketch } from './FullPageSketch.js';
+export { default as SketchController } from './SketchController.js';
