--- src/components/page/index.ts
@@ -1,2 +1,0 @@
-import Page from './Page.js';
-export default Page;
