+++ src/index.ts
@@ -1,2 +1,3 @@
 export * from './sketch/index.js';
+export * from './components/index.js';
 export { default as App } from './App.js';
